moved header to separate control
