+++ VSRAD.Package/DebugVisualizer/SliceVisualizer/SliceVisualizerHeaderControl.xaml
@@ -1,0 +1,13 @@
+﻿<UserControl x:Class="VSRAD.Package.DebugVisualizer.SliceVisualizer.SliceVisualizerHeaderControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:VSRAD.Package.DebugVisualizer.SliceVisualizer"
+             mc:Ignorable="d" 
+             d:DesignHeight="30" d:DesignWidth="800">
+    <StackPanel Orientation="Horizontal">
+        <TextBlock Text="Watch: "/>
+        <ComboBox ItemsSource="{Binding Watches}" SelectionChanged="NewWatchSelected" Height="25" Width="70"/>
+    </StackPanel>
+</UserControl>
